serialization_api

Trait ApiSerializations

pub trait ApiSerializations {
    type Address;
    type RewardClaimInput;
    type RewardBalance;
    type RewardAccountQueryData;
    type RewardBalances;
    type RewardMerkleTreeData;

    // Required methods
    fn deserialize_address(&self, s: &str) -> Result<Self::Address>;
    fn serialize_reward_claim_input(
        &self,
        address: &str,
        value: &Self::RewardClaimInput,
    ) -> Result<RewardClaimInput>;
    fn serialize_reward_balance(
        &self,
        value: &Self::RewardBalance,
    ) -> Result<RewardBalance>;
    fn serialize_reward_account_query_data(
        &self,
        value: &Self::RewardAccountQueryData,
    ) -> Result<RewardAccountQueryDataV2>;
    fn serialize_reward_balances(
        &self,
        value: &Self::RewardBalances,
    ) -> Result<RewardBalances>;
    fn serialize_reward_merkle_tree_data(
        &self,
        value: &Self::RewardMerkleTreeData,
    ) -> Result<RewardMerkleTreeV2Data>;
}

Trait for converting between implementation types and proto serialization types

Implementations define their own internal types for addresses and rewards data, then provide conversions to/from the proto types for API serialization.

Required Associated Types

type Address

Address type used by the implementation

type RewardClaimInput

Reward claim input type

type RewardBalance

Reward balance type

type RewardAccountQueryData

Reward account query data type (balance + proof)

type RewardBalances

Paginated reward balances type

type RewardMerkleTreeData

Reward merkle tree snapshot data type

Required Methods

fn deserialize_address(&self, s: &str) -> Result<Self::Address>

Deserialize an address string from a proto request into the implementation’s Address type

fn serialize_reward_claim_input( &self, address: &str, value: &Self::RewardClaimInput, ) -> Result<RewardClaimInput>

Serialize implementation’s RewardClaimInput to proto RewardClaimInput

Takes the original address string since the internal type may not contain it

fn serialize_reward_balance( &self, value: &Self::RewardBalance, ) -> Result<RewardBalance>

Serialize implementation’s RewardBalance to proto RewardBalance

fn serialize_reward_account_query_data( &self, value: &Self::RewardAccountQueryData, ) -> Result<RewardAccountQueryDataV2>

Serialize implementation’s RewardAccountQueryData to proto RewardAccountQueryDataV2

fn serialize_reward_balances( &self, value: &Self::RewardBalances, ) -> Result<RewardBalances>

Serialize implementation’s RewardBalances to proto RewardBalances

fn serialize_reward_merkle_tree_data( &self, value: &Self::RewardMerkleTreeData, ) -> Result<RewardMerkleTreeV2Data>

Serialize implementation’s RewardMerkleTreeData to proto RewardMerkleTreeV2Data